+++ src/lib/contacts/contacts.ts
@@ -5,8 +5,8 @@ import * as io from "../interfaces/http-io";
 import { Contact } from "../types/contact";
 import { Invite } from "../types/invite";
 import { Url } from "../types/url";
+import { getContacts } from "./api/get-contacts";
 import { $GetInvitesResult, GetInvitesResult } from "./api/get-invites";
-import { $GetUserResult, GetUserResult } from "./api/get-user";
 import * as contactsUrl from "./contacts-url";
 
 export interface ContactsInterface {
@@ -56,32 +56,17 @@ export class ContactsService {
     } catch (err) {
       throw new Incident(err, "UnexpectedResponseBody", {body: response.body});
     }
-    const result: GetInvitesResult = $GetInvitesResult.readJson(parsed);
+
+    let result: GetInvitesResult;
+    try {
+      result = $GetInvitesResult.readJson(parsed);
+    } catch (err) {
+      throw new Incident(err, "UnexpectedResult", {body: parsed});
+    }
     return result.inviteList;
   }
 
   async getContacts(apiContext: Context): Promise<Contact[]> {
-    const url: Url = contactsUrl.formatUser(apiContext.username);
-    const request: io.GetOptions = {
-      uri: url,
-      queryString: {page_size: "100", reason: "default"},
-      cookies: apiContext.cookies,
-      headers: {
-        "X-Skypetoken": apiContext.skypeToken.value,
-      },
-    };
-    const response: io.Response = await this.httpIo.get(request);
-    if (response.statusCode !== 200) {
-      UnexpectedHttpStatusError.create(response, new Set([200]), request);
-    }
-    let parsed: any;
-    try {
-      parsed = JSON.parse(response.body);
-    } catch (err) {
-      throw new Incident(err, "UnexpectedResponseBody", {body: response.body});
-    }
-    const result: GetUserResult = $GetUserResult.readJson(parsed);
-    // TODO: Expose the other properties.
-    return result.contacts;
+    return getContacts(this.httpIo, apiContext);
   }
 }